chore: move diff and cast to pkg
