+++ pkg/parser/token.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/go-errors/errors"
 	"github.com/spf13/viper"
-	"github.com/supabase/cli/internal/utils/cast"
+	"github.com/supabase/cli/pkg/cast"
 )
 
 // Equal to `startBufSize` from `bufio/scan.go`
